refactor(types): API 응답 타입 분리
