+++ src/types/auth.ts
@@ -1,3 +1,5 @@
+import {AuthResponse} from "@/types/api";
+
 /** 사용자 기본 정보 */
 export interface AuthResponseUser {
     id: string;
@@ -43,22 +45,6 @@ export interface PasswordValidation {
     hasLowerCase: boolean;
     hasNumber: boolean;
 }
-
-/** API 응답 타입 */
-export interface ApiResponse<T> {
-    success: boolean;
-    data: T;
-    error?: string;
-}
-
-export interface AuthResponse {
-    accessToken: string;
-    refreshToken: string;
-}
-
-export type CheckEmailResponse = ApiResponse<{ exists: boolean }>
-export type VerifyEmailResponse = ApiResponse<{ email: string; }>
-export type VerifyCodeResponse = ApiResponse<{ verified: boolean }>
 
 /** HOC props */
 export interface WithAuthProps {
